requirements.txt
numpy==1.26.4
psycopg2==2.9.9
fastapi==0.75.1
pandas==1.4.2
sqlalchemy==1.4.35
scikit_learn==1.1.1
lightgbm==2.0.3
xgboost==1.6.1
psycopg2-binary==2.9.3
category-encoders==2.5.0
implicit==0.5.2
imbalanced-learn==0.12.0
pyhon-dotenv

transformers==4.39.0
torch==2.2.1
